test: break out common
